turbo_taxi_booking_system/requirements.txt
pillow
psycopg
tkcalendar
tkintermapview
haversine
pydantic